+++ A32NX/html_ui/Pages/VCockpit/Instruments/A320_Neo/Map/A32NX_MapInstrument.js
@@ -1,3 +1,21 @@
+/*
+ * A32NX
+ * Copyright (C) 2020-2021 FlyByWire Simulations and its contributors
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
 class SmartIterator {
     constructor() {
         this._minReturned = NaN;
@@ -257,6 +275,9 @@ class MapInstrument extends ISvgMapRootElement {
             this.showFlightPlan = false;
             if (newValue === "true") {
                 this.showFlightPlan = true;
+                this.updateFlightPlanVisibility();
+            }
+            if (newValue == "false") {
                 this.updateFlightPlanVisibility();
             }
         } else if (lowercaseName === "hide-flightplan-if-bushtrip") {